Address undefined behaviour in agg
```
    y=14725117.312413637) at deps/agg/include/agg_rasterizer_outline_aa.h:127
    y=14725117.312413637, cmd=1) at deps/agg/include/agg_rasterizer_outline_aa.h:145
    this=0x7ffe29ad4ef8, vs=..., path_id=0) at deps/agg/include/agg_rasterizer_outline_aa.h:175
    at include/mapnik/vertex_converters.hpp:332
```

Then after addressing iround, I got a crash from a different cast:
```
    at deps/agg/include/agg_rasterizer_outline_aa.h:51
    at deps/agg/include/agg_vertex_sequence.h:85
    at deps/agg/include/agg_rasterizer_outline_aa.h:121
    y=-1921023.6493361052) at deps/agg/include/agg_rasterizer_outline_aa.h:133
```

It kept coming:
```
    at deps/agg/include/agg_rasterizer_outline_aa.h:377
    y=44374331484.875168, cmd=79) at deps/agg/include/agg_rasterizer_outline_aa.h:151
    this=0x7ffcdf3595c8, vs=..., path_id=0) at deps/agg/include/agg_rasterizer_outline_aa.h:175
```

And on:
```
    at deps/agg/include/agg_renderer_outline_image.h:987
    at deps/agg/include/agg_rasterizer_outline_aa.h:248
    at deps/agg/include/agg_rasterizer_outline_aa.h:384
    y=44374331484.875168, cmd=79) at deps/agg/include/agg_rasterizer_outline_aa.h:151
```

and on:
```
    at deps/agg/include/agg_rasterizer_outline_aa.h:304
    at deps/agg/include/agg_rasterizer_outline_aa.h:384
    y=44374331484.875168, cmd=79) at deps/agg/include/agg_rasterizer_outline_aa.h:151
```

and on:
```
    this=0x560703a4a818, ras=..., x1=65536, y1=2147483647, x2=65536, y2=-491781926, f1=2, f2=8)
    at deps/agg/include/agg_rasterizer_sl_clip.h:180
180                         tx1 = x1 + Conv::mul_div(m_clip_box.y2-y1, x2-x1, y2-y1);
(gdb) bt
    this=0x560703a4a818, ras=..., x1=65536, y1=2147483647, x2=65536, y2=-491781926, f1=2, f2=8)
    at deps/agg/include/agg_rasterizer_sl_clip.h:180
    this=0x560703a4a818, ras=..., x2=2147483647, y2=-491781926) at deps/agg/include/agg_rasterizer_sl_clip.h:246
    x=11231485.450064059, y=-1921023.1493361052) at deps/agg/include/agg_rasterizer_scanline_aa.h:403
    x=11231485.450064059, y=-1921023.1493361052, cmd=2) at deps/agg/include/agg_rasterizer_scanline_aa.h:420
    at deps/agg/include/agg_rasterizer_scanline_aa.h:198
    geom=..., proc=...) at include/mapnik/vertex_converters.hpp:332
```

and finally:
```
    this=0x560492a53518, ras=..., x1=65536, y1=2147483647, x2=65536, y2=-491781926, f1=2, f2=8)
    at deps/agg/include/agg_rasterizer_sl_clip.h:183
183                         tx1 = x1 + Conv::mul_div(m_clip_box.y2-y1, x2-x1, y2-y1);
(gdb) bt
    this=0x560492a53518, ras=..., x1=65536, y1=2147483647, x2=65536, y2=-491781926, f1=2, f2=8)
    at deps/agg/include/agg_rasterizer_sl_clip.h:183
    this=0x560492a53518, ras=..., x2=2147483647, y2=-491781926) at deps/agg/include/agg_rasterizer_sl_clip.h:255
    x=11231485.450064059, y=-1921023.1493361052) at deps/agg/include/agg_rasterizer_scanline_aa.h:403
    x=11231485.450064059, y=-1921023.1493361052, cmd=2) at deps/agg/include/agg_rasterizer_scanline_aa.h:420
    at deps/agg/include/agg_rasterizer_scanline_aa.h:198
    (geom=..., proc=...) at include/mapnik/vertex_converters.hpp:332
```

+++ deps/agg/include/agg_rasterizer_outline_aa.h
@@ -47,8 +47,8 @@ namespace agg
 
         bool operator () (const line_aa_vertex& val)
         {
-            double dx = val.x - x;
-            double dy = val.y - y;
+            double dx = static_cast<double>(val.x) - x;
+            double dy = static_cast<double>(val.y) - y;
             return (len = uround(sqrt(dx * dx + dy * dy))) > 
                    (line_subpixel_scale + line_subpixel_scale / 2);
         }
